pyproject.toml
[project]
name = "polymonitor"
version = "0.1.0"
description = "Polymarket personal trading monitor"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "python-telegram-bot>=21.0",
    "pyyaml>=6.0",
    "apscheduler>=3.10,<4.0",
    "fastapi>=0.115",
    "uvicorn>=0.34",
]

[project.scripts]
polymonitor = "src.main:main"
